lesson_30/requirements.txt
psycopg2-binary==2.9.9
pytest==8.3.3
allure-pytest==2.13.5